+++ hata/discord/webhook/webhook.py
@@ -175,7 +175,7 @@ class Webhook(WebhookBase):
         
         banner_color = data.get('banner_color', None)
         if (banner_color is not None):
-            banner_color = Color(banner_color[1:])
+            banner_color = Color(banner_color[1:], 16)
         self.banner_color = banner_color
     
     @classmethod